refactor(sidebar): refact sidebar module
