+++ src/index.js
@@ -1,10 +1,9 @@
 import './assets/app.scss'
-const d3 = require('d3')
+import renderSidebar from './modules/sidebar'
 
 document.addEventListener('DOMContentLoaded', () => {
   readData((data) => {
-    const fields = getDataFields(data)
-    renderSideBar(fields)
+    renderSidebar(data)
   })
 })
 
@@ -23,24 +22,4 @@ function readData(callback) {
   if(callback) {
     callback(data)
   }
-}
-
-function getDataFields(data) {
-  return Object.keys(data[0])
-}
-
-function renderSideBar(fields) {
-  const sidebar = d3.select(".sidebar");
-  sidebar
-    .selectAll("input")
-    .data(fields)
-    .enter()
-    .append("label")
-      .attr("for", (d) => d)
-      .text((d) => d)
-    .append("input")
-      .attr("checked", true)
-      .attr("type", "checkbox")
-      .attr("id", (d) => d)
-      .on("click", (d) => console.log(`Pressed ${d}`))
 }
